Moves a class to a different package
